06-ia_model/requirements.txt
# Para reconocimiento facial
face_recognition==1.3.0  # Requiere dlib
dlib==19.24.2  # Versión compatible con face_recognition
opencv-python==4.8.0  # Para procesamiento de imágenes

# Para YOLOv8
ultralytics==8.0.0  # Incluye torch (PyTorch)
torch==2.0.0  # Requerido por ultralytics (para GPU usa torch==2.0.0+cu118)

# Para MQTT
paho-mqtt==1.6.1

# Utilerías
numpy==1.24.0